Added event handling
We also have event binding which we use to handle events raised from the DOM keystrokes mouse movements
clicks and so on.
So here we have a simple button and we want to handle the click event of this button.
So instead of square brackets use parenthesis and here we are at the name of the event like click then
we bind this to a method in our component.
So let's call this on save.
Now here I create this method on say and simply log a message in a console console that log button was
clicked.
So here's our button click.
Now let's take a look at the cons. OK here's the message now sometimes we need to get access to the
event object that was raised in the event handler for example with mouse movements the event object
will tell us that x and y position if you want to get access to that event object we need to add that
as a parameter here.
So dollar event and then when calling this method you also pass dollar event dollar event object is
something known to angular.
Now in this case we are dealing with a DOM object.
So this dom or event object will represent a dom event a standard Dom event that you have seen in vanilla
Javascript.

So I can or on Save method.
Let's lock this event dollar event same in the browser.
So one more time click OK here's our daughter event objects all these properties you see here are part
of the standard Dom event objects so we have the movement x and y offset X and Y and many other properties.
Now all the dummy events bubble up the DOM tree unless a handler along the way prevents further bubbling.
This is just a standard event propagation mechanism in DOM and it's nothing specific about angular.
For example let me wrap this button with a div like this.

Now this Div I also want to handle the click event so practices click and I want to bind this to a
different method.
So let's call this on div clicked
let's add this method here on div click.
And here I want to log a different message div was clicked.
Now let's see what happens.
So when we click this button we get two messages in the console.
The first one is the handler for the Click event of the button.
And the second is from the handler of the Click event of the div.
So this is what we call event bubbling and event bubbles up the DOM tree.

So to extend this example if we have another div or another element that contained this div and we handle
the click event on that element our event object will bubble love and hate that target handler.
How can we stop event bubbling.
Well here in on say we can call $event.stopPropagation()

Again this is a standard method we have seen before in vanilla Javascript.
So when we call this this event will not bubble up.
In other words it's not going to hit the second handler a straight save like in the browser.
Click now look in the console.
We only have one message that's coming from the handler of the Click event of the button.