Added template variable
Are containing from our last example let's imagine we want to get the value that was typed into this input field.
How can we do this.
There are two ways.
One way is to use the event object.
So once again we past this event object here at this parameter to our method and I'm going to remove

this message and simply log daughter event.

Now this is a standard event object in dumb.

So it has a target property that references this input field and from here we can get the value.

Let's try this back in the browser.
So if I type and get her and press Enter NOW in the console we have angular unifor now in Angular we
have another way to solve the same problem.

Instead of passing this object around we can declare a viable in our template that references this input field.
So here I declare a variable like this.
Let's imagine this input fill is for capturing the user's e-mail.
So I declare a viable using this hash tag or the pound sign.
And this is the name of our valuable.

We call this a template variable and this references this input field.
Now instead of passing this event object around we can pass e-mail dot value OK.
And then as the parameter to this method we can simply receive the actual e-mail address and log in

on the case like this.
Let's try it so me@xample.com.
Enter.
And here in the con. We got the e-mail address.
So in a lot of cases you can use template variables to simplify your code.